requirements.txt
netmiko>=4.3.0
pynetbox>=7.1.0
PyYAML>=6.0
urllib3>=2.0
